backout yopypi

+++ bootstrap.sh
@@ -81,9 +81,7 @@ VPYTHON_DIR="$PRE$BASE_DIR"
 VIRTUALENV="virtualenv-1.6.4"
 VIRTUALENV_TARBALL="${VIRTUALENV}.tar.gz"
 PIP="./${VPYTHON_DIR}/bin/pip"
-#PIP_OPTS="--index-url=http://127.0.0.1:8080/simple"
 PIP_OPTS="--use-mirrors"
-YOPYPI="./${VPYTHON_DIR}/bin/yopypi-cli"
 
 # only install if we dont already exist
 if [ ! -d $VPYTHON_DIR ]
@@ -113,10 +111,10 @@ then
     # create a virtual python in the current directory
     $TARGET_PYTHON ${CACHE}/$VIRTUALENV/build/lib*/virtualenv.py --no-site-packages $VPYTHON_DIR
 
-    ${PIP} install yopypi
-    ${YOPYPI} start
-    ${PIP} install ${PIP_OPTS} -r requirements.txt
-    ${YOPYPI} stop
+    if [ -f "requirements.txt" ]
+    then
+        ${PIP} install ${PIP_OPTS} -r requirements.txt
+    fi
 
     # hack activate to set some environment we need
     echo "PROJECT_DIRECTORY=`pwd`;" >>  $VPYTHON_DIR/bin/activate
